Lift giant-planet wire set into p9-core
