+++ crates/p9-core/src/initial_conditions/mod.rs
@@ -1,2 +1,3 @@
+pub mod giant_planets;
 pub mod planets;
 pub mod scattered_disk;
